Move config folder for root directory
