+++ src/auth/handlers/redirect/auth-redirect-handler.js
@@ -1,6 +1,6 @@
 const generateID = require('../../../utils/generate-id')
 const stringify = require('../../../utils/stringify')
-const { spotify } = require('../../../config/environment')
+const { spotify } = require('../../../../config/environment')
 
 module.exports = (request, response) => {
   const stateValue = generateID()
