+++ src/server.js
@@ -1,4 +1,4 @@
-const { port, webClient } = require('./config/environment.js')
+const { port, webClient } = require('../config/environment.js')
 const express = require('express')
 const cors = require('cors')
 const authRouter = require('./auth/router/auth-router')
